+++ internal/consts/consts.go
@@ -1066,6 +1066,14 @@ const (
 	AuthMethodAzure    = "azure"
 
 	/*
+		Azure auth_type values
+	*/
+	AuthTypeRootCreds = "root_creds"
+	AuthTypePluginWIF = "plugin_wif"
+	AuthTypeMSI       = "msi"
+	AuthTypeAKSWI     = "aks_wi"
+
+	/*
 		misc. path related constants
 	*/
 	PathDelim        = "/"