+++ vault/resource_terraform_cloud_secret_backend.go
@@ -205,6 +205,9 @@ func terraformCloudSecretBackendRead(ctx context.Context, d *schema.ResourceData
 	if err := readMount(ctx, d, meta, true, true); err != nil {
 		return diag.FromErr(err)
 	}
+	if d.Id() == "" {
+		return nil
+	}
 
 	log.Printf("[DEBUG] Reading %s from Vault", configPath)
 	secret, err := client.Logical().ReadWithContext(ctx, configPath)
@@ -350,6 +353,11 @@ func terraformCloudSecretBackendDelete(ctx context.Context, d *schema.ResourceDa
 	log.Printf("[DEBUG] Unmounting Terraform Cloud backend %q", backend)
 	err := client.Sys().UnmountWithContext(ctx, backend)
 	if err != nil {
+		if util.Is404(err) {
+			log.Printf("[DEBUG] Terraform Cloud backend %q not found, removing from state", backend)
+			d.SetId("")
+			return nil
+		}
 		return diag.Errorf("Error unmounting Terraform Cloud backend from %q: %s", backend, err)
 	}
 	log.Printf("[DEBUG] Unmounted Terraform Cloud backend %q", backend)